move to simple file treatment plan routes
